Refactor types package
