+++ lib/src/main/kotlin/me/tb/cashuclient/types/Keyset.kt
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
  */
  
-package me.tb.cashuclient
+package me.tb.cashuclient.types
 
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
