+++ lib/src/main/kotlin/me/tb/cashuclient/types/PreRequestBundle.kt
@@ -1,7 +1,5 @@
 package me.tb.cashuclient.types
 
-import me.tb.cashuclient.KeysetId
-
 public interface PreRequestBundle {
     public val blindingDataItems: List<BlindingData>
     public val keysetId: KeysetId
