+++ lib/src/main/kotlin/me/tb/cashuclient/types/Secret.kt
@@ -3,7 +3,9 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the ./LICENSE file.
  */
  
-package me.tb.cashuclient
+package me.tb.cashuclient.types
+
+import me.tb.cashuclient.randomBytes
 
 /**
  * Secret used to generate a token. This is the x in NUT-00, the bytes we'll use in the [hashToCurve] function
